Moved moscaSetup file to tests directory
